온도 타입 수정

+++ src/main/java/project/app/apns_server/modules/vo/AppInfoVo.java
@@ -48,7 +48,7 @@ public class AppInfoVo {
         return apnsId.hashCode();
     }
 
-    public void updateCurrTemp(double currTemp) {
+    public void updateCurrTemp(long currTemp) {
         this.temp = currTemp;
     }
 }
